Make beep work on a Raspberry Pi (using wiringPi)
This patch adds support for using `beep` on a Raspberry Pi,
using an old BIOS speaker or piezo buzzer attached to its
GPIO PWM pin.

When *not* running as root, ugly `system()` calls are used
to call the `gpio` binary included with `wiringPi`. Note that
this is significantly slower, as the binary needs to be called
5 times for each tone! You might be better off running `beep`
as `suid`.

The original Makefile has been replaced with a Makefile.in to
accommodate the conditional compilation with `wiringPi` support.
Build with `./configure && make`.