beep.c: rewrite to use drivers; allow all devnames
Completely uses driver infrastructure now.

This means it is now easy to choose at compile time which
drivers (console, evdev) are to be used and in which sequence.

Stops checking the given device name for sanity.

beep refuses to run with elevated privileges (setuid or sudo)
any more, so this does not allow any priviledge escalation via
the --device argument.