Use sig_atomic_t for signal handlers to communicate with main thread
The signal handlers now do not call anything any more.  To compensate
for that, all loops in the main thread check for the global_abort flag
set by the signal handlers, and after every nanosleep (which may be
interrupted by a signal).

So if any signal handler is triggered, the main thread should abort
very quickly.