Only check if running under sudo when running as root
There is no priviledge escalation potential if e.g. root uses sudo
to drop priviledges and run processes as a non-root user which in
turn call `beep`.

The purpose of `beep`'s check for running under sudo is to
prevent `beep` from running with old setups and old packages
which used to set up `beep` setuid-root or as an alias to
`sudo beep` to run `beep` as root which was mostly required
with the old console API.

The check still serves that purpose.

Fixes: #5