log_init fix: if argv[0] ends in slash, keep default progname
We now check for

    if ('\0' != *post_last_slash)

instead of the intended

    if (*post_last_slash)

which we had mistyped as the semantically very different

    if (post_last_slash)

which made it possible to set progname (printed in the log messages)
to argv[0] if argv[0] ended in a '/'. Now we keep progname the
intended default value in that case.

Also, we now keep the default value for progname in case argv[0]
should be empty.

The impact of those two issues were the wrong string being printed
in log messages, but still a string, and thus this is not security
relevant.